// apps/farcaster-bot/__generated__/ees/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigDecimal: { input: any; output: any; }
  BigInt: { input: any; output: any; }
  Bytes: { input: any; output: any; }
  /**
   * 8 bytes signed integer
   *
   */
  Int8: { input: any; output: any; }
};

export type Account = {
  __typename?: 'Account';
  id: Scalars['Bytes']['output'];
  receivedDonations: Array<Donation>;
  receivedEndorsements: Array<Endorsement>;
  sentDonations: Array<Donation>;
  sentEndorsements: Array<Endorsement>;
  totalDonationsReceived: Scalars['BigInt']['output'];
  totalDonationsSent: Scalars['BigInt']['output'];
  totalEndorsementsReceived: Scalars['BigInt']['output'];
  totalEndorsementsSent: Scalars['BigInt']['output'];
  withdrawals: Array<Withdrawal>;
};
// ...
export type AggregatedInformation = {
  __typename?: 'AggregatedInformation';
  donationAmount: Scalars['BigInt']['output'];
  endorsementCount: Scalars['BigInt']['output'];
  from: Account;
  id: Scalars['Bytes']['output'];
  to: Account;
};
// ...
export type Donation = {
  __typename?: 'Donation';
  amount: Scalars['BigInt']['output'];
  from: Account;
  id: Scalars['Bytes']['output'];
  to: Account;
};
// ...
export type Endorsement = {
  __typename?: 'Endorsement';
  donationAmount: Scalars['BigInt']['output'];
  easUid: Scalars['Bytes']['output'];
  endorsementType: Scalars['String']['output'];
  from: Account;
  id: Scalars['Bytes']['output'];
  to: Account;
};
// ...
export type FeeWithdrawal = {
  __typename?: 'FeeWithdrawal';
  address: Scalars['Bytes']['output'];
  amount: Scalars['BigInt']['output'];
  id: Scalars['Bytes']['output'];
  withdrawalType: Scalars['String']['output'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  account?: Maybe<Account>;
  accounts: Array<Account>;
  aggregatedInformation?: Maybe<AggregatedInformation>;
  aggregatedInformations: Array<AggregatedInformation>;
  donation?: Maybe<Donation>;
  donations: Array<Donation>;
  endorsement?: Maybe<Endorsement>;
  endorsements: Array<Endorsement>;
  feeWithdrawal?: Maybe<FeeWithdrawal>;
  feeWithdrawals: Array<FeeWithdrawal>;
  withdrawal?: Maybe<Withdrawal>;
  withdrawals: Array<Withdrawal>;
};
// ...
export type Withdrawal = {
  __typename?: 'Withdrawal';
  address: Account;
  amount: Scalars['BigInt']['output'];
  id: Scalars['Bytes']['output'];
};
// ...
export type _Meta_ = {
  __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars['String']['output'];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars['Boolean']['output'];
};

export enum _SubgraphErrorPolicy_ {
  /** Data will be returned even if the subgraph has indexing errors */
  Allow = 'allow',
  /** If the subgraph has indexing errors, data will be omitted. The default. */
  Deny = 'deny'
}

export type GetTopEndorseesQueryVariables = Exact<{
  first: Scalars['Int']['input'];
  skip: Scalars['Int']['input'];
}>;


export type GetTopEndorseesQuery = { __typename?: 'Query', accounts: Array<{ __typename?: 'Account', id: any, totalEndorsementsReceived: any }> };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}

export const GetTopEndorseesDocument = new TypedDocumentString(`
    query GetTopEndorsees($first: Int!, $skip: Int!) {
  accounts(
    orderBy: totalEndorsementsReceived
    orderDirection: desc
    first: $first
    skip: $skip
  ) {
    id
    totalEndorsementsReceived
  }
}
    `) as unknown as TypedDocumentString<GetTopEndorseesQuery, GetTopEndorseesQueryVariables>;